requirements.txt
anyio==4.12.1
APScheduler==3.11.2
certifi==2026.2.25
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
python-dotenv==1.2.2
python-telegram-bot==22.6
PyYAML==6.0.3
tzlocal==5.3.1
yt-dlp==2026.3.13
blinker==1.9.0
click==8.3.1
gunicorn==25.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.6
requests==2.32.3
instaloader
